samples/agent/adk/mcp-app-sample/pyproject.toml
# Copyright 2026 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "a2ui-mcp-app-sample"
version = "0.1.0"
description = "Sample MCP App with 2-Way Communication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "google-adk>=1.28.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "fastapi>=0.123.10",
    "uvicorn>=0.40.0",
    "starlette>=0.50.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
